Update README.md
#waitforit, 2023, interactive websites

A series of pure front-end re-creations of popular application loading screens, spinners, and progress bars presented as two interactive websites.  Any interaction on the first website leads to another loading screen. The second site displays every loading screen included in the work as a randomly generated grid of iframes.

#waitforit examines how fundamental human needs like food, safety, socializing, sex, mobility, etc. are transactionalized on the web through instant convenience apps and online consumer platforms.  The loading screens, spinners, and progress bars unfold in an online drama of human needs and desires, the things worth waiting for, and hint at the deliberately obscured complexities of network economics and online control waiting on the other end.

Spinners and progress bars are the potential for a moment, a promise, they are the Internet holding it’s breath.  Comprised of the most basic of shapes, usually lines and circles:

In Japan, there was a tradition of composing a poem on your deathbed, called a jisei. One final reflection on the totality of your life, a gift to your family, and a distillation of a complex existence into something beautiful.
In 1769, the dying poet Shisue was asked by his students to compose a jisei, and with his brush, the master drew a single circle. No words, just a single line to represent the void, the universe in totality, and enlightenment.

Sometimes the simplest mark is the most powerful.

– Excerpted from: Reed Enger, "Onement, I," in Obelisk Art History.